Fix first review errors *.html
* <ul class="two-columns__main-text">
Here the text paragraphs are not an enum,
so a div tag can be used to group them, instead of ul.

* <li class="two-columns__paragraph">
Text blocks li must be marked with a paragraph tag p.

* <h4 class="table__heading">86 миллиардов</h4>
* <h4 class="table__heading">1</h4>
Headings to the third level.

* Узнать→
 Подробнее →
 <p class="footer__author">© 2022. Medvedeva</p>
Use mnemonics.
For example, the mnemonic &rarr; for →
Mnemonic &copy; for ©.

* <a href="#" class="footer__column-link">
Add target="_blank" attribute to external links